+++ karma.conf.js
@@ -22,8 +22,9 @@ module.exports = function(config) {
       'bower_components/restangular/dist/restangular.js',
       'bower_components/localforage/dist/localforage.js',
       'bower_components/angular-localForage/dist/angular-localForage.js',
+      'src/index.js',
       'src/**/*.js',
-      'tests/**/*.spec.js',
+      'tests/**/*.js',
       '.tmp/templates.js'
     ],